refactor: move findWinner method
